+++ src/Checkout/Payment/Service/OrderExecuteService.php
@@ -11,6 +11,7 @@ use Psr\Log\LoggerInterface;
 use Shopware\Core\Checkout\Order\Aggregate\OrderTransaction\OrderTransactionStateHandler;
 use Shopware\Core\Framework\Context;
 use Shopware\Core\Framework\Log\Package;
+use Shopware\Core\System\StateMachine\Exception\IllegalTransitionException;
 use Shopware\PayPalSDK\Struct\ConstantsV2;
 use Shopware\PayPalSDK\Struct\V2\Order as PayPalOrder;
 use Shopware\PayPalSDK\Struct\V2\Order\PurchaseUnit\Payments;
@@ -87,6 +88,15 @@ class OrderExecuteService
                 return false;
             }
 
+            if ($capture->getStatus() === ConstantsV2::ORDER_CAPTURE_PENDING) {
+                try {
+                    $this->orderTransactionStateHandler->process($transactionId, $context);
+                } catch (IllegalTransitionException) {
+                }
+
+                return true;
+            }
+
             if ($capture->getStatus() === ConstantsV2::ORDER_CAPTURE_COMPLETED) {
                 $this->orderTransactionStateHandler->paid($transactionId, $context);
 
@@ -104,6 +114,15 @@ class OrderExecuteService
         $authorization = $this->getPayments($order, $salesChannelId, $refetch)?->getAuthorizations()?->first();
         if ($authorization === null) {
             return false;
+        }
+
+        if ($authorization->getStatus() === ConstantsV2::ORDER_AUTHORIZATION_PENDING) {
+            try {
+                $this->orderTransactionStateHandler->process($transactionId, $context);
+            } catch (IllegalTransitionException) {
+            }
+
+            return true;
         }
 
         if ($authorization->getStatus() === ConstantsV2::ORDER_AUTHORIZATION_CREATED) {