+++ tests/Mock/PayPalSDK/MockRequestHandler.php
@@ -44,6 +44,7 @@ use Swag\PayPal\Test\Mock\PayPal\Client\_fixtures\V2\CaptureAuthorization;
 use Swag\PayPal\Test\Mock\PayPal\Client\_fixtures\V2\CaptureOrderAPM;
 use Swag\PayPal\Test\Mock\PayPal\Client\_fixtures\V2\CaptureOrderCapture;
 use Swag\PayPal\Test\Mock\PayPal\Client\_fixtures\V2\CaptureOrderDeclined;
+use Swag\PayPal\Test\Mock\PayPal\Client\_fixtures\V2\CaptureOrderPending;
 use Swag\PayPal\Test\Mock\PayPal\Client\_fixtures\V2\CreateOrderAPM;
 use Swag\PayPal\Test\Mock\PayPal\Client\_fixtures\V2\CreateOrderCapture;
 use Swag\PayPal\Test\Mock\PayPal\Client\_fixtures\V2\CreateOrderPUI;
@@ -225,6 +226,10 @@ class MockRequestHandler
 
             if (\mb_substr($resourceUri, -17) === GetOrderCaptureLiabilityShiftUnknown::ID) {
                 return $this->createResponse(SymResponse::HTTP_OK, GetOrderCaptureLiabilityShiftUnknown::get());
+            }
+
+            if (\mb_substr($resourceUri, -17) === CaptureOrderPending::ID) {
+                return $this->createResponse(SymResponse::HTTP_OK, CaptureOrderPending::get());
             }
 
             $orderCapture = GetOrderCapture::get();